requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
youtube-transcript-api